+++ CHANGELOG.md
@@ -11,6 +11,7 @@ but instead follows [Break Versioning](https://www.taoensso.com/break-versioning
 ### Added
 
 * `scribe.config` - Simple config loading
+* `scribe.string/strip-ansi-color` - Strip ANSI color sequences
 
 ## [1.0.0] - 2024-08-19